and finally no errors time to test

+++ src/apps/molresponse/Plot_VTK.cc
@@ -46,8 +46,8 @@ void do_vtk_plots(World &world, int npt_plot, double L, int plotlo, int plothi,
   std::vector<Vector<double, 3>> coords = molecule.get_all_coords_vec();
 
   // Write the data
-  int Natoms = molecule.natom();
-  for (int i = 0; i < Natoms; i++) {
+size_t   Natoms = molecule.natom();
+  for (size_t    i = 0; i < Natoms; i++) {
     std::fprintf(f, "%5s   %16.12f %16.12f %16.12f\n",
                  atomic_number_to_symbol(molecule.get_atom_number(i)).c_str(),
                  coords[i][0], coords[i][1], coords[i][2]);
@@ -60,7 +60,7 @@ void do_vtk_plots(World &world, int npt_plot, double L, int plotlo, int plothi,
   real_function_3d rho = real_factory_3d(world);
 
   // Plot each orbital requested
-  for (int i = plotlo; i < plothi; i++) {
+  for (size_t    i = plotlo; i < plothi; i++) {
     // Add to total
     rho += densities[i];
